fix(client): decode a fractional job progress and pin the raw render
An integer progress field rejects the whole job envelope, so one
fractional percentage fails a healthy poll. The render test gained the
branch where a body carries a code and no message, which must keep the
raw form a caller matches on.

+++ internal/client/polling.go
@@ -18,17 +18,19 @@ type JobStatusSingularResponse struct {
 	Data    JobResponse `json:"data"`
 }
 type JobResponse struct {
-	JobID              string `json:"jobId"`
-	Stage              string `json:"stage"`
-	IsCompleted        bool   `json:"isCompleted"`
-	IsTerminal         bool   `json:"isTerminal"`
-	HasFailed          bool   `json:"hasFailed"`
-	ProgressPercentage int    `json:"progressPercentage"`
-	Message            string `json:"message"`
-	ErrorMessage       string `json:"errorMessage"`
-	ResourceId         string `json:"resourceId"`
-	ResourceName       string `json:"resourceName"`
-	ResourceType       string `json:"resourceType"`
+	JobID       string `json:"jobId"`
+	Stage       string `json:"stage"`
+	IsCompleted bool   `json:"isCompleted"`
+	IsTerminal  bool   `json:"isTerminal"`
+	HasFailed   bool   `json:"hasFailed"`
+	// The field is a float, because a percentage the platform computes can
+	// arrive fractional. An integer field rejects the whole envelope.
+	ProgressPercentage float64 `json:"progressPercentage"`
+	Message            string  `json:"message"`
+	ErrorMessage       string  `json:"errorMessage"`
+	ResourceId         string  `json:"resourceId"`
+	ResourceName       string  `json:"resourceName"`
+	ResourceType       string  `json:"resourceType"`
 }
 
 // JobStageCancelled is the stage an operator cancel leaves a job in. It is
